rename: 폴더명 대소문자 수정
